fix(build): resolve -Wall warnings from stricter compile flags
- sim_data.h: widen PhaseTiming burstGap/mouseDur/idleDur fields from
  uint16_t to uint32_t (break/meeting modes need durations >65.5s)
- display.cpp: bump batStr[8] → [16] and posInd[8] → [10] to silence
  -Wformat-truncation for theoretical int range of %d format args

+++ sim_data.h
@@ -16,10 +16,10 @@ struct ProfileWeights {
 struct PhaseTiming {
   uint8_t  burstKeysMin, burstKeysMax;
   uint16_t interKeyMinMs, interKeyMaxMs;
-  uint16_t burstGapMinMs, burstGapMaxMs;
+  uint32_t burstGapMinMs, burstGapMaxMs;
   uint16_t keyHoldMinMs, keyHoldMaxMs;
-  uint16_t mouseDurMinMs, mouseDurMaxMs;
-  uint16_t idleDurMinMs, idleDurMaxMs;
+  uint32_t mouseDurMinMs, mouseDurMaxMs;
+  uint32_t idleDurMinMs, idleDurMaxMs;
 };
 
 struct WorkModeDef {
